+++ shell/platform/linux/BUILD.gn
@@ -81,6 +81,7 @@ source_set("flutter_linux_sources") {
              "fl_dart_project_private.h",
              "fl_engine_private.h",
              "fl_keyboard_handler.h",
+             "fl_keyboard_manager.h",
              "fl_keyboard_pending_event.h",
              "fl_keyboard_view_delegate.h",
              "fl_key_event.h",
@@ -118,6 +119,7 @@ source_set("flutter_linux_sources") {
     "fl_key_responder.cc",
     "fl_keyboard_handler.cc",
     "fl_keyboard_layout.cc",
+    "fl_keyboard_manager.cc",
     "fl_keyboard_pending_event.cc",
     "fl_keyboard_view_delegate.cc",
     "fl_message_codec.cc",
@@ -217,6 +219,7 @@ executable("flutter_linux_unittests") {
     "fl_key_embedder_responder_test.cc",
     "fl_keyboard_handler_test.cc",
     "fl_keyboard_layout_test.cc",
+    "fl_keyboard_manager_test.cc",
     "fl_message_codec_test.cc",
     "fl_method_channel_test.cc",
     "fl_method_codec_test.cc",
